Add RangeIndexedInts.
